feat: implement ISpanParsable<T>
For time being the impl. is private, until we can commit to a complete API change (deprecating IbanParser). The behavior is nuanced, because the IbanParser accepts a custom registry whereas this impl. uses the default registry. We either change this via a static settable registry, or don't bother, but this is a choice that must be made before we can make this public and phase out IbanParser.